added weighting to umap rotation
omg, i think the reason I'm not getting the plots i like is that the plot function is constraining it to a certain limit, instead of diong that automatically. in linear ena, the scaling is *important*, but in nonlinear, it ain't, so long as you can read it. simple fix?